Add recycler view to add friend dialogue
Tapping a method to add a friend will change the dialogue
to show an edit text with a hint specific to the method as
well as an empty recycler view to be filled with search results.
If choosing from facebook the users list of facebook friends who
have signed into Snaption will populate the results list. Searching
has no effect on the results list and must be implemented