Pull the users friends to populate the friend list
When a user logs in their list of friends is saved locally via realm.
Then this information is also pulled down everytime the friends fragment
is created to get any new information. The data saved on realm can be
utilized later if we want to have offline capabilities. Additionally
the information being sent back from the server lacks full names so
the friends list looks a little empty without them.