Add the ability to remove friends
Users can now swipe left to remove their friends on the friends
list. A confirmation dialog will appear once they swipe and if
they confirm the user is deleted, if not it is moved back in place.
Note that the delete endpoint needs some very specific syntax to
allow for body parameters